# Conflicts:
#	ConfluenceOtherworld/src/main/resources/confluence.mixins.json
#	EnemyBanner

+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/common/hook/ManaConsumeHook.java
@@ -3,9 +3,8 @@ package org.confluence.mod.common.hook;
 import com.xiaohunao.equipment_benediction.common.hook.IHook;
 import com.xiaohunao.equipment_benediction.common.interfaces.IBenediction;
 import net.minecraft.world.item.ItemStack;
-
-import java.util.function.IntSupplier;
+import org.confluence.mod.util.FloatSupplier;
 
 public interface ManaConsumeHook extends IHook {
-    IntSupplier onManaConsume(IBenediction owner, ItemStack itemStack, IntSupplier original);
+    FloatSupplier onManaConsume(IBenediction owner, ItemStack itemStack, FloatSupplier original);
 }
