+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/common/item/common/EverBeneficialItem.java
@@ -55,7 +55,7 @@ public class EverBeneficialItem extends TooltipItem {
     });
     public static final Beneficial AMBROSIA = new Beneficial(Confluence.asResource("ambrosia"), EverBeneficial::setAmbrosiaUsed, (id, player, everBeneficial, isRespawn) -> {
         int value = TCUtils.getAccessoriesValue(player, TCItems.RIGHT$CLICK$DELAY$SUBSTRACTOR);
-        PacketDistributor.sendToPlayer(player, new RightClickSubtractorPacketS2C(Math.min(value + 1, 4)));
+        PacketDistributor.sendToPlayer(player, new RightClickSubtractorPacketS2C((byte) Math.min(value + 1, 4)));
         AttributeInstance attributeInstance = player.getAttributes().getInstance(Attributes.BLOCK_BREAK_SPEED);
         if (attributeInstance == null) return;
         attributeInstance.addOrReplacePermanentModifier(new AttributeModifier(id, 0.05, AttributeModifier.Operation.ADD_MULTIPLIED_TOTAL));
